+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use the official `sailfish` crate instead of the forked version, since the official `sailfish` crate now supports removing newlines in the rendered HTML files.
 - Upgrade Rust to v1.92.
-
+- Upgrade to dependencies to their latest versions.
 
 ## [0.3.0]